Fixed HTTP2 protocol

+++ httpparser/lib/http2.cc
@@ -11,7 +11,7 @@ std::ostream& operator<<(std::ostream& os, const header_request &header_request)
     const http_header_request &http11request = header_request;
 
     os << "Stream Identifier: " << header_request.stream_identifier << std::endl
-        << "Weight: " << header_request.weight << std::endl;
+        << "Weight: " << (unsigned int)header_request.weight << std::endl;
     
     if (header_request.error != frame::error_t::NO_ERROR) {
         os << "Error: " << header_request.error << std::endl;
